refactor: improve files structure
