+++ components/BirdInfo/BirdInfo.module.scss
@@ -1,4 +1,4 @@
-@import './variables/all.scss';
+@import '../../styles/variables/all.scss';
 
 .birdsInfoContainer {
   display: flex;
