+++ components/Header/Header.module.scss
@@ -1,4 +1,4 @@
-@import './variables/all.scss';
+@import '../../styles/variables/all.scss';
 
 .container {
   width: 100%;
@@ -23,7 +23,7 @@
 }
 
 .logo {
-  background-image: url("../public/svg/logo.svg");
+  background-image: url("../../public/svg/logo.svg");
   width: 200px;
   height: 55px;
   background-size: contain;
