+++ components/Nav/Nav.module.scss
@@ -1,4 +1,4 @@
-@import './variables/all.scss';
+@import '../../styles/variables/all.scss';
 
 .pagination {
   display: flex;
