+++ constants/imageSize.js
@@ -1,0 +1,6 @@
+const imageSize = {
+  width: 200,
+  height: 155
+}
+
+export default imageSize
