+++ constants/levels.js
@@ -1,0 +1,10 @@
+const levels = {
+  introduction: 'Разминка',
+  sparrows: 'Воробьиные',
+  forestbirds: 'Лесные птицы',
+  songbirds: 'Певчие птицы',
+  predatorbirds: 'Хищные птицы',
+  seabirds: 'Морские птицы'
+}
+
+export default levels
