+++ pages/404.js
@@ -1,6 +1,5 @@
-import styles from '../styles/404.module.scss';
+import Page404 from '../scenes/Page404/Page404';
 import Head from 'next/head';
-import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 
@@ -18,10 +17,7 @@ const Error = () => {
       <Head>
         <title>error</title>
       </Head>
-      <div className={styles.container}>
-        <h2 className={styles.title}>Something is going wrong...</h2>
-        <Image src="/img/404.png" width={270} height={170} alt="404" />
-      </div>
+      <Page404 />
     </>
   );
 };
