+++ pages/cards/card.module.scss
@@ -1,4 +1,4 @@
-@import './variables/all.scss';
+@import '../../styles/variables/all.scss';
 
 .container {
   width: 100%;
