+++ scenes/Page404/Page404.module.scss
@@ -1,5 +1,3 @@
-@import './variables/all.scss';
-
 .container {
   display: flex;
   flex-direction: column;
